Get giá tiền với thông tin dịch vụ từ appoinemntType

+++ src/main/java/com/KoiHealthService/Koi/demo/entity/AppointmentType.java
@@ -17,6 +17,5 @@ public class AppointmentType {
     @GeneratedValue(strategy = GenerationType.UUID)
     String appointmentTypeId;
     String appointmentService;
-    Double price;
-
+    Long price;
 }
